서버에서 온 결과에 따른 스크립트 수정

+++ Assets/SimulFactory/Script/Ui/Battle/UiPlayerControlPanel.cs
@@ -49,7 +49,7 @@ namespace SimulFactory.Ui.Battle
         /// <param name="rspType"></param>
         private void ButtonClicked(Define.ROCK_SCISSOR_PAPER rspType)
         {
-            BattleManager.GetInstance().ButtonClicked(rspType);
+            BattleManager.GetInstance().ButtonClicked((int)rspType);
         }
     }
 }
